+++ dc-cudami-admin/src/main/java/de/digitalcollections/cudami/admin/controller/identifiable/entity/EventsAPIController.java
@@ -1,0 +1,54 @@
+package de.digitalcollections.cudami.admin.controller.identifiable.entity;
+
+import de.digitalcollections.cudami.admin.business.api.service.exceptions.ServiceException;
+import de.digitalcollections.cudami.admin.controller.ParameterHelper;
+import de.digitalcollections.cudami.admin.controller.identifiable.AbstractIdentifiablesController;
+import de.digitalcollections.cudami.admin.model.bootstraptable.BTResponse;
+import de.digitalcollections.cudami.admin.util.LanguageSortingHelper;
+import de.digitalcollections.cudami.client.CudamiClient;
+import de.digitalcollections.cudami.client.identifiable.entity.CudamiEventsClient;
+import de.digitalcollections.model.exception.TechnicalException;
+import de.digitalcollections.model.identifiable.entity.Event;
+import de.digitalcollections.model.list.paging.PageResponse;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import java.util.UUID;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+/** Controller for all public "Events" endpoints (API). */
+@RestController
+public class EventsAPIController
+    extends AbstractIdentifiablesController<Event, CudamiEventsClient> {
+
+  public EventsAPIController(LanguageSortingHelper languageSortingHelper, CudamiClient client) {
+    super(client.forEvents(), languageSortingHelper, client.forLocales());
+  }
+
+  /*
+  Used in templates/events/list.html
+  */
+  @SuppressFBWarnings
+  @GetMapping("/api/events")
+  @ResponseBody
+  public BTResponse<Event> find(
+      @RequestParam(name = "offset", required = false, defaultValue = "0") int offset,
+      @RequestParam(name = "limit", required = false, defaultValue = "1") int limit,
+      @RequestParam(name = "search", required = false) String searchTerm,
+      @RequestParam(name = "sort", required = false, defaultValue = "label") String sort,
+      @RequestParam(name = "order", required = false, defaultValue = "asc") String order,
+      @RequestParam(name = "dataLanguage", required = false) String dataLanguage)
+      throws TechnicalException, ServiceException {
+    PageResponse<Event> pageResponse =
+        super.find(localeService, service, offset, limit, searchTerm, sort, order, dataLanguage);
+    return new BTResponse<>(pageResponse);
+  }
+
+  @GetMapping("/api/events/{uuid:" + ParameterHelper.UUID_PATTERN + "}")
+  @ResponseBody
+  public Event getByUuid(@PathVariable UUID uuid) throws TechnicalException {
+    return service.getByUuid(uuid);
+  }
+}