+++ CHANGELOG.md
@@ -17,7 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Added dropdown to select number of rows per page in table lists on `Articles`, `DigitalObjects`, `Collections`
 - Added status storing of new tables (navigating back to list shows list in last status)
 - Added highlighting of searched term in result lists in new table lists on `Articles`, `DigitalObjects`, `Collections`
-- Add backend for `Event`
+- Add backend and GUI for `Event`
 
 ### Changed